Fix bug in normalize_r Single-precision temp variable could result in ry > 1 after normalization, causing a domain error in asin(ry).

+++ source/cryxtels.cpp
@@ -1787,7 +1787,7 @@ void extra_stop_extra ()
 // among other things, that would allow us to easily pass vectors to functions and return them
 void normalize_r ()
 {
-    float norm = 1.0 / std::sqrt(rx * rx + ry * ry + rz * rz);
+    double norm = 1.0 / std::sqrt(rx * rx + ry * ry + rz * rz);
     rx *= norm;
     ry *= norm;
     rz *= norm;
